--- a/Makefile
+++ b/Makefile
@@ -2,7 +2,6 @@ GOFMT ?= gofmt -s
 GOFILES := $(shell find . -name "*.go" -type f -not -path "./vendor/*")
 RELEASE_OPTIONS := -ldflags "-s -w -X github.com/goby-lang/goby/vm.DefaultLibPath=${GOBY_LIBPATH}" -tags release
 TEST_OPTIONS := -ldflags "-s -w"
-BENCHMARK_OPTIONS := -run '^$$' -bench '.' -benchmem -benchtime 2s
 
 .PHONY: fmt
 fmt:
@@ -23,15 +22,3 @@ test:
 .PHONY: clean
 clean:
 	go clean .
-
-.PHONY: update_benchmarks
-update_benchmarks:
-	go test $(BENCHMARK_OPTIONS) ./... > current_benchmarks
-
-.PHONY: compare_benchmarks
-compare_benchmarks: 
-	go test $(BENCHMARK_OPTIONS) ./... > .tmp_benchmarks 
-	benchcmp current_benchmarks .tmp_benchmarks
-
-
-