Adds a simple ruby script to compare benchmarks of two git checkouts.

Saving off the results of a benchmarks doesn't make a lot of sense because we currently can't run the benchmarks on the same machine. This will allow us to compare benchmarks on a per diff basis apples to apples. 

Example output.

```
benchmarking 2fd6637efac2bf8c371f020ab1e95aa0f7606873
benchmarking 59b7a7b303e2b0f6e783e34360c3b352192f1a57
universal.x86_64-darwin17
benchmark                         old ns/op     new ns/op     delta
BenchmarkBasicMath/add-8          5734          5651          -1.45%
BenchmarkBasicMath/subtract-8     5651          5700          +0.87%
BenchmarkBasicMath/multiply-8     5676          5636          -0.70%
BenchmarkBasicMath/divide-8       5771          5699          -1.25%

benchmark                         old allocs     new allocs     delta
BenchmarkBasicMath/add-8          78             78             +0.00%
BenchmarkBasicMath/subtract-8     78             78             +0.00%
BenchmarkBasicMath/multiply-8     78             78             +0.00%
BenchmarkBasicMath/divide-8       78             78             +0.00%

benchmark                         old bytes     new bytes     delta
BenchmarkBasicMath/add-8          3792          3792          +0.00%
BenchmarkBasicMath/subtract-8     3792          3792          +0.00%
BenchmarkBasicMath/multiply-8     3792          3792          +0.00%
BenchmarkBasicMath/divide-8       3792          3792          +0.00%
Switched to branch 'update-benchmars'
```